+++ services/web/server/src/simcore_service_webserver/projects/_crud_api_read.py
