+++ services/web/server/tests/conftest.py
@@ -24,7 +24,7 @@ from servicelib.aiohttp.long_running_tasks.server import TaskStatus
 from servicelib.json_serialization import json_dumps
 from simcore_service_webserver.application_settings_utils import convert_to_environ_vars
 from simcore_service_webserver.db.models import UserRole
-from simcore_service_webserver.projects._crud_create_utils import (
+from simcore_service_webserver.projects._crud_api_create import (
     OVERRIDABLE_DOCUMENT_KEYS,
 )
 from simcore_service_webserver.projects.models import ProjectDict
