+++ services/web/server/tests/integration/01/test_garbage_collection.py
@@ -40,7 +40,7 @@ from simcore_service_webserver.groups.api import (
     list_user_groups,
 )
 from simcore_service_webserver.login.plugin import setup_login
-from simcore_service_webserver.projects._crud_delete_utils import get_scheduled_tasks
+from simcore_service_webserver.projects._crud_api_delete import get_scheduled_tasks
 from simcore_service_webserver.projects.models import ProjectDict
 from simcore_service_webserver.projects.plugin import setup_projects
 from simcore_service_webserver.resource_manager.plugin import setup_resource_manager
