+++ services/web/server/tests/integration/01/test_project_workflow.py
@@ -136,14 +136,14 @@ async def storage_subsystem_mock(mocker: MockerFixture):
         )
 
     mock = mocker.patch(
-        "simcore_service_webserver.projects._crud_create_utils.copy_data_folders_from_project",
+        "simcore_service_webserver.projects._crud_api_create.copy_data_folders_from_project",
         autospec=True,
         side_effect=_mock_copy_data_from_project,
     )
 
     # requests storage to delete data
     mock1 = mocker.patch(
-        "simcore_service_webserver.projects._crud_delete_utils.delete_data_folders_of_project",
+        "simcore_service_webserver.projects._crud_api_delete.delete_data_folders_of_project",
         return_value="",
     )
     return mock, mock1
@@ -171,7 +171,7 @@ def catalog_subsystem_mock(
         return services_in_project
 
     for namespace in (
-        "simcore_service_webserver.projects._crud_read_utils.get_services_for_user_in_product",
+        "simcore_service_webserver.projects._crud_api_read.get_services_for_user_in_product",
         "simcore_service_webserver.projects._crud_handlers.get_services_for_user_in_product",
     ):
         mock = mocker.patch(
