+++ services/web/server/tests/unit/with_dbs/01/studies_dispatcher/test_studies_dispatcher_handlers.py
@@ -279,7 +279,7 @@ def catalog_subsystem_mock(mocker: MockerFixture) -> None:
     ]
 
     mock = mocker.patch(
-        "simcore_service_webserver.projects._crud_read_utils.get_services_for_user_in_product",
+        "simcore_service_webserver.projects._crud_api_read.get_services_for_user_in_product",
         autospec=True,
     )
