+++ services/web/server/tests/unit/with_dbs/02/test_projects_handlers__list_search.py
@@ -35,7 +35,7 @@ def standard_user_role() -> tuple[str, tuple[UserRole, ExpectedResponse]]:
 @pytest.fixture
 def mock_catalog_api_get_services_for_user_in_product(mocker: MockerFixture):
     mocker.patch(
-        "simcore_service_webserver.projects._crud_read_utils.get_services_for_user_in_product",
+        "simcore_service_webserver.projects._crud_api_read.get_services_for_user_in_product",
         spec=True,
         return_value=[],
     )
